pyproject.toml
[project]
name = "langgraph_app"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "uvicorn>=0.35.0",
    "sse-starlette>=2.1.0,<2.2.0",
    "uvloop>=0.18.0",
    "httptools>=0.5.0",
    "jsonschema-rs>=0.20.0",
    "structlog>=24.1.0",
    "cloudpickle>=3.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "anyio>=4.7.0",
    "fastmcp>=2.10.4",
    "grpcio>=1.73.1",
    "pydantic>=2.11.7",
    "qdrant-client>=1.14.3",
    "openai>=1.95.1",
    "uvicorn>=0.35.0",
    "fastembed>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.30.0",
    "aiohttp>=3.13.1",
    "jinja2>=3.1.6",
    "aiofiles>=25.1.0",
    "types-aiofiles>=25.1.0.20251011",
    "langgraph-checkpoint>=3.0.1",
    "langchain-core>=1.0.1",
    "langgraph>=1.0.1",
    "langgraph-sdk>=0.2.9",
    "langchain-openai>=1.0.2",
    "langsmith>=0.4.38",
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.1.12",
    "langgraph-api==0.7.28",
    "langgraph-checkpoint-postgres>=3.0.1",
    "psycopg[binary,pool]>=3.3.0",
    "google-auth>=2.47.0",
    "google-api-python-client>=2.188.0",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
    "pytest-asyncio>=1.2.0",
    "langgraph-cli[inmem]>=0.2.8",
]

[tool.uv]
# всегда подтягивать dev-зависимости
default-groups = ["dev"]

[tool.ruff]
line-length = 88

lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # первая строка docstring — повелительное наклонение
    "T201", # принты
    "UP",   # pyupgrade
]

lint.ignore = [
    "UP006",
    "UP007",
    "UP035", # разрешено import из typing_extensions
    "D417",  # не требуем полных docstring для всех аргументов
    "E501",  # не проверяем длину строки
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
plugins = ["pydantic.mypy"] # Включает плагин Pydantic для корректной типизации моделей и автогенерации сигнатур конструкторов

ignore_missing_imports = true # глобально игнорирует отсутствие type stubs для сторонних библиотек,

follow_imports = "silent" # Не выводить предупреждения о типах в импортируемых модулях (ускоряет проверку и уменьшает шум)
warn_redundant_casts = true # Предупреждать о ненужных операциях приведения типа (например, cast, которые не влияют на вывод типов)
warn_unused_ignores = true # Предупреждать, если используется # type: ignore, которые фактически не нужны (чистота кода)
disallow_any_generics = true # Запретить обобщенные типы с неявным any (например, List без указания типа элементов)
check_untyped_defs = true # Проверять функции без аннотаций типов (строже, помогает найти ошибки в неаннотированном коде)
no_implicit_reexport = true # Не экспортировать типы импорты из __init__.py автоматически (контролирует импортируемые типы)
disallow_untyped_defs = true # Требовать аннотаций типов у всех функций и методов (повышает строгость проверки)

[tool.pydantic-mypy]
init_forbid_extra = true # Запретить передачу в конструктор моделей Pydantic полей, не описанных явно в модели
init_typed = true # Генерировать типизированные сигнатуры __init__ на основе типов атрибутов модели (типизация конструктора)
warn_required_dynamic_aliases = true # Предупреждать, если используются алиасы или динамические имена полей без явной типизации
